Generify BaseListEntry to use the "curiously recurring template" pattern
This allows it to specify the types of the input functions in terms of the inheriting class (eventually). The `self()` method is a type-safe `this`, since `this` doesn't know that it should be the same type as the `SELF` parameter.

This ultimately cleans up a lot of ugly casting and prevents some simple mistakes when accessing the entry from inside a function argument.