Introduce a "Dependency" class
To allow for having multiple dependencies and non-boolean dependencies, lets store dependencies in the form of a class.

This allows any one entry to depend on multiple other entries. **All** dependencies must be met for the entry to be enabled.

A dependency can be set to "hide when disabled", if one of the unmet dependencies has this set, the entry is not shown at all.

A dependency can have multiple conditions defined, if **any** condition is met the dependency is satisfied.