Ensure tooltips stay on screen
When the tooltip is wider than half the screen width, it often ends up
partially off-screen.

Let's ensure the tooltip always wraps to less than `screen.width / 2`.

Additionally, lets render the tooltip above the cursor, instead of
in-line with it. This allows the user to more easily read what they're
hovering over without the tooltip getting in the way.