Refactor DependencyManager's registry
An entry's field was being overwritten by parent fields as `DependencyManager.register()` got called multiple times.

Instead of keeping track of the field, we can simplify things be maintaining a list of dependency annotations that have been registered with the entry.

This also means we can get rid of the separate `additionalDependencies` map and just have a list of annotations in the main `EntryRecord`.

This greatly simplifies `buildDependencies()`

This fixes outer fields being unable to depend on children of their siblings.