Relative i18n references
It is now possible to depend on other fields using a relative i18n reference.

By starting the reference with a `.`, the referencing field's parent key will be inserted as a prefix.

It's also possible to start a reference with multiple `.`s. For each `.`, one "generation" of parent is removed from the prefix, allowing you to reference "cousin" objects too.