Separate hide behaviour from Dependency interface
Dependencies shouldn't need to know how dependent guis will behave when the dependency is met/unmet.

Instead, we can define an "enable if" and a "show if" dependency on each gui.

For autoconfig, we can handle that with separate annotations; `@EnableIf` and `@ShowIf` with corresponding `*Group` annotations.

In `DependencyManager` we can simply map the annotation to a `DependencyRecord` and use this instead of the old `@DependsOn` annotation.

In `FieldBuilder` we can split `withDependency()` into `setEnabledIf()` and `setShownIf()`.