DependencyBuilder split into Start* and Finish*
`StartDependencyBuilder` is used to start the building process and select a `ConfigEntry` to depend on.

`FinishDependencyBuilder` is used to finalise the process, implementors may provide additional implementation-specific methods that only make sense for their specific type.

`ConfigEntryDependencyBuilder` and `DependencyGroupBuilder` are both `FinishDependencyBuilder` implementations.

Non-type specific methods (such as `generateTooltip()`) should be declared on the most generic interface. Any values set on the `StartDependencyBuilder` should be handed-over to the `FinishDependencyBuilder`.