Enable/Display config entries using dependencies
Allows users to define dependency `Requirement`s as lambda functions, or generate them via factory methods.

A standard requirement controls whether a Config Entry is "enabled" or "disabled". When "disabled" the Config Entry is greyed-out and cannot be interacted with.

A "display" requirement controls whether a Config Entry appears _at all_ in the GUI. It's still technically present, but won't be rendered or take up any space in the list GUI.

A hidden Config Entry is also disabled automatically.